packages/tools/pyproject.toml
[project]
name = "terramod-schema"
version = "0.1.0"
description = "Terraform schema to multi-language YAML converter"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
]

[project.scripts]
terramod-schema = "terramod_schema.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/terramod_schema"]
